+++ lib/cocoapods-core/specification/consumer.rb
@@ -88,9 +88,20 @@ module Pod
       spec_attr_accessor :compiler_flags
 
       # @return [Hash{String => String}] the xcconfig flags for the current
-      #         specification.
+      #         specification for the pod target.
       #
-      spec_attr_accessor :xcconfig
+      def pod_target_xcconfig
+        attr = Specification::DSL.attributes[:pod_target_xcconfig]
+        merge_values(attr, value_for_attribute(:xcconfig), value_for_attribute(:pod_target_xcconfig))
+      end
+
+      # @return [Hash{String => String}] the xcconfig flags for the current
+      #         specification for the user target.
+      #
+      def user_target_xcconfig
+        attr = Specification::DSL.attributes[:user_target_xcconfig]
+        merge_values(attr, value_for_attribute(:xcconfig), value_for_attribute(:user_target_xcconfig))
+      end
 
       # @return [String] The contents of the prefix header.
       #