+++ lib/cocoapods-core/specification/dsl/deprecations.rb
@@ -9,6 +9,17 @@ module Pod
           CoreUI.warn "[#{self}] `preferred_dependency` has been renamed "\
             'to `default_subspecs`.'
         end
+
+        DSL.attribute :xcconfig,
+                      :container => Hash,
+                      :inherited => true
+
+        def xcconfig=(value)
+          self.pod_target_xcconfig = value
+          self.user_target_xcconfig = value
+          CoreUI.warn "[#{self}] `xcconfig` has been split into "\
+            '`pod_target_xcconfig` and `user_target_xcconfig`.'
+        end
       end
     end
   end